backend/requirements.txt
fastapi>=0.115.0
pydantic>=2.9.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
psycopg[binary,pool]==3.3.3
redis==5.0.1
python-jose[cryptography]==3.3.0
cryptography==41.0.5
pydantic-settings==2.1.0
email-validator
python-multipart==0.0.9
slowapi>=0.1.9
pyotp>=2.9.0
qrcode[pil]>=7.4.2
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.25.0
aiosqlite>=0.19.0
bcrypt
greenlet